Refactored food-up
